pyproject.toml
[tool.poetry]
name = "ress-smalltalk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ress_smalltalk"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
aiohttp = "^3.8.3"
python-telegram-bot = {version = "^20.0a4", allow-prereleases = true}
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
oitnb = "^0.2.2"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
